rename python to pip
