+++ pkg/pip/parse_test.go
@@ -1,4 +1,4 @@
-package python
+package pip
 
 import (
 	"os"
