+++ pkg/pip/testdata/requirements_no_version.txt
