+++ pkg/pip/testdata/requirements_operator.txt
